kernel: sched: Introduce N-EMS for 4.19 Kernels The Not Exynos mobile scheduler.
Its the EMS (Exynos Mobile Scheduler) (Kind of) Ported for Snapdragon devices.

It also introduces energy_aware & energy_step, i tried making
energy_aware work but failed, but the rest seems to be alright,
energy_step is as it is since i havent tried messing with it.
**Keep in mind this is All Experimental For now

Default EMS Kconfig text:
It is designed to secure the limits of energy aware scheduler. This option provides
features such as independent boosting functinos such as on-time migration,
and prefer_perf and enhanced prefer_idle that work in conjunction with
SCHEDTUNE.

Author: João Batista <sonicgames772@gmail.com>